Make post form be more like Twitter with live client-side validations
This commit removes most of the element of the form just leaving the
content input which becomes a textarea.

It also add default values for likes count, reposts count and
username.

There's also validation added to ensure the content is provided and is
at least two characters and 250 at most.

LiveView being connected using a socket to the server, each change to
the form state triggers a validation so the user knows in real-time if
the content provided is valid.